Added a new example: The "TestEchoClientServer"
This example is 1 dense .cpp file consisting of a singla main() function
that basically illustrates how to use the library (but it's dense code
so not for the faint of heart).  I used it while testing/developing the
library as a little sanity checker.  I figured I would release it with
the library since others may find it useful.

Also in this commit: moved the user contributed examples to examples/contrib