Updated codebase to work with Qt6 out of the box
The only change needed was to the sample TestEchoClientServer.cpp
application to not use deprecated QRegExp but instead use the newer
QRegularExpression (which is supported in Qt5 as well).

Since most of the codebase relies simply on low-level types such as
QByteArray and QTcpSocket -- and since these are largely unchanged from
Qt5 to Qt6 -- there doesn't appear to be any issue using WebSocket.cpp
in Qt6.